file: renamed files for services, i.e., REST and WebSocket
