+++ src/brokers/mexc/http_gateway.py
@@ -4,7 +4,7 @@ import json
 
 # Custom libraries
 from src.infrastructure.logging.set_logger import operation_logger
-from src.brokers.base.rest_sdk import RestService
+from src.brokers.base.http_sdk import RestService
 
 
 class MexcFutureGateway(RestService):
